+++ abc-bank/IAccount.cs
@@ -7,10 +7,10 @@ namespace abc_bank
     {
         Account.AccountType Type { get; }
         IList<Transaction> Transactions { get; }
-        void Deposit(Double amount);
-        void Withdraw(Double amount);
-        Double SumTransactions();
-        Double SumTransactions(DateTime toDate);
-        Double InterestEarned();
+        void Deposit(Decimal amount);
+        void Withdraw(Decimal amount);
+        Decimal SumTransactions();
+        Decimal SumTransactions(DateTime toDate);
+        Decimal InterestEarned();
     }
 }
